pyproject.toml
[project]
name = "faiss"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
]
